refactor: Extractor class
