+++ src/tab/scraper/scraper.js
@@ -1,7 +1,7 @@
 import {domainExportChecker} from "../checker/domainChecker";
 import {logWelcome} from "../utils/consoleMessages";
 import {setFormatRules} from "./ruler/ruler";
-import {extract} from "./extractor/extractor";
+import {extract} from "./extractor/extract";
 import {exportContent} from "../exporter/exporter";
 import {clickOnListElt} from "../uiEnhancer/phind/interact";
